Increase resolution artwork in Plex section
As discussed in #105 (sorry it took me  a while to get to this) I've increased the resolution of artwork within the Plex page so that they look better on high resolution displays too.  Didn't bump too high to keep cache size down so they're hardly "retina" images but to the eye they look sharp unless you were to super zoom right in on one.