Improvements to Sonarr module
Added episode numbering to Sonarr calendar view.

On Sonarr dash module episode titles are always so truncated that they aren't really useful, so I've replaced the title with the epsode number and added a tooltip next to the episode number which gives the overview of that episode.  I think it makes the Sonarr dash module more informative.

Minor typo fixes.